% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iSEEindex-package.R
\docType{package}
\name{iSEElogin-package}
\alias{iSEElogin-package}
\title{iSEElogin: iSEE extension for a landing page presenting a login form}
\description{
This package provides a landing page that protects the main app behind
a login page.
The main functionality of this package is to define a custom landing page
that presents users with a login form.
Users cannot create new accounts.
Instead, the landing page must be accompanied by a file encrypting
a predefined set of usernames and passwords authorised to log in.
}
\seealso{
Useful links:
\itemize{
\item \url{https://github.com/iSEE/iSEElogin}
\item Report bugs at \url{https://support.bioconductor.org/t/iSEElogin}
}
}
\author{
\strong{Maintainer}: Kevin Rue-Albrecht \email{kevinrue67@gmail.com} (\href{https://orcid.org/0000-0003-3899-3872}{ORCID})

}
\keyword{internal}
