% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iSEEindex.R
\name{iSEElogin}
\alias{iSEElogin}
\title{iSEElogin App}
\usage{
iSEElogin(app.title = NULL)
}
\arguments{
\item{app.title}{Character string to specify the desired title to be displayed in the main window of the dashboard.
Defaults to \code{NULL}, which displays the versions of the \code{iSEElogin} and \code{iSEE} packages.}
}
\value{
An \code{\link[iSEE:iSEE]{iSEE::iSEE()}} app with a custom landing page presenting users with a log in form.
}
\description{
Generate an \pkg{iSEE} app that includes a landing page forcing users
to log in using any pair of authorised username and password
from a predefined set prepared by the app maintainer.
}
\examples{
##
# iSEElogin ----
##

app <- iSEElogin()

if (interactive()) {
    shiny::runApp(app, port = 1234)
}
}
\author{
Kevin Rue-Albrecht
}
